requirements.txt
pandas
transformers
torch
accelerate
streamlit
streamlit-extras
watchdog
beautifulsoup4
tqdm
requests
datasets
browser_cookie3
lmstudio
groq
google-genai
cerebras_cloud_sdk
